% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{DeriveData}
\alias{DeriveData}
\title{Derive demographics dataset - this would be a function that is called when CRON is running}
\usage{
DeriveData()
}
\description{
Derive demographics dataset - this would be a function that is called when CRON is running
}
