% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{GetDataset}
\alias{GetDataset}
\title{Get demographics dataset}
\usage{
GetDataset(platform = NULL)
}
\value{
The demographics dataset with age group breakdown and new insvestor flag data.table
}
\description{
Get demographics dataset
}
