% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{GetInceptionYearDistribution}
\alias{GetInceptionYearDistribution}
\title{Get inception year distribution}
\usage{
GetInceptionYearDistribution(platform = NULL)
}
\value{
The inception year distribution data.table
}
\description{
Get inception year distribution
}
